CPLTMInfo

+++ app.verify/A.Core/10.TuringMachine/ICPLTMInfo.cs
@@ -17,7 +17,7 @@ namespace Core
 
     uint PathLength { get; }
 
-    List<int> KTapeLRSubseq();
+    List<long> KTapeLRSubseq();
   }
 }
